requirements.txt
dash==2.17.0
dash_daq==0.5.0
dash_bootstrap_components==1.5.0
dash_core_components==2.0.0
dash_extensions==1.0.16
dash_html_components==2.0.0
dash_slicer==0.3.1
dash_vtk==0.0.9
dipy==1.7.0
diskcache==5.6.3
dash[diskcache]
Flask==3.0.3
kaleido==0.2.1
fury==0.8.0
ipywidgets==8.1.3
kaggle==1.6.14
matplotlib==3.6.3
nibabel==4.0.2
numpy==1.23.5
pandas==2.2.2
patsy==0.5.6
Pillow==9.3.0
plotly==5.20.0
pycountry==24.6.1
python-dotenv==1.0.1
requests==2.28.2
scikit-learn==1.2.2
scilpy==1.5.0
scipy==1.9.3
seaborn==0.13.2
scikit-image==0.22.0
statsmodels==0.13.5
tqdm==4.66.2
umap_learn==0.5.5
vtk==9.2.6
XlsxWriter==3.2.0